\name{bamboo}
\alias{bamboo}
\title{Just Give it a Shot}
\description{
Just give it a shot and you either will or will not regret it.
}
\usage{
bamboo(x = "world")
}
\arguments{
  \item{x}{The recipient of the fancy message generated by this very function.}
}
\value{
The function returns nothing, invisibly.
}
\examples{
bamboo()
bamboo("Karen")
}
