% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State code in FARS data}

\item{year}{A numeric year in yyyy format}
}
\value{
A map with the location of every accident
in the specified state during the specified
year.
}
\description{
Accepts a FARS state code and a year and returns
a map where every accident in that state for
that year is represented by a point. Will throw
an error for invalid state codes.
x = longitude and y = latitute
}
\examples{
\dontrun{
fars_map_state("01", 2013)
}
}
