% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string which contain the path
to the file you want to read.}
}
\value{
A tibble containing the file data
}
\description{
Given a path to a csv file, return a tibble
which contains the file data. Will stop
if the file does not exist.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
