% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector of years to read}
}
\value{
A list of tibbles each with FARS
casualty data for the corresponding years.
}
\description{
Accepts a list or vector of years. Returns a
list of tibbles with the casualty data for each
year. Will throw an error for an invalid year.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014))
}
}
