% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector of years to read}
}
\value{
A tibble whose rows are months.
Each column represents one year and the values
are the number of accidents in that month
and year
}
\description{
Fetches the accident summary for each month for
each of the years given as input.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014))
}
}
