\name{rpackagedemo-package}
\alias{rpackagedemo-package}
\alias{rpackagedemo}
\docType{package}
\title{
\packageTitle{rpackagedemo}
}
\description{
\packageDescription{rpackagedemo}
}
\author{
\packageAuthor{rpackagedemo}

Maintainer: \packageMaintainer{rpackagedemo}
}
\keyword{ package }
\seealso{
\code{\link{demo}}
}
\examples{
demo()
demo("Karen")
}
