% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Title}
\usage{
square(x)
}
\arguments{
\item{x}{A character vector with one element.}
}
\value{
split What to split on
}
\description{
Title
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
