\name{demo}
\alias{demo}
\title{Does a Demo}
\description{
Just give it a shot and you won't regret it.
}
\usage{
demo(x = "world")
}
\arguments{
  \item{x}{The recipient of the fancy message generated by this very function.}
}
\value{
The function returns nothing, invisibly.
}
\examples{
demo()
demo("Karen")
}
